Introducing AbstractMultiMap
